+++ src/RFC4180FieldTest.php
@@ -54,9 +54,9 @@ final class RFC4180FieldTest extends TestCase
         $csv = Writer::createFromPath('php://temp');
         RFC4180Field::addTo($csv);
         self::assertContains(RFC4180Field::getFiltername(), stream_get_filters());
-        $csv->setNewline("\r\n");
+        $csv->setEndOfLine("\r\n");
         $csv->insertOne($record);
-        self::assertSame($expected, $csv->getContent());
+        self::assertSame($expected, $csv->toString());
     }
 
     public static function bugsProvider(): array