+++ CHANGELOG.md
@@ -6,13 +6,17 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Added
 
-- `TabulatDataReader::select` (implemented on the `Reader` and the `ResultSet` object)
-- `FragmentParser` to implement [RFC7111](https://www.rfc-editor.org/rfc/rfc7111)
+- `TabulatDataReader::select`
+- `TabulatDataReader::matching`
+- `TabulatDataReader::firstMatching` 
+- `TabulatDataReader::firstOrFailMatching`
+- `FragmentFinder` to implement [RFC7111](https://www.rfc-editor.org/rfc/rfc7111)
 
 ### Deprecated
 
 - Using the `$header` argument on `Statement::process` is deprecated and will be removed in
 the next version. Use `TabularDataReader::getRecords` on the returned value instead.
+It's usage will trigger a `E_USER_DEPRECATED` call.
 
 ### Fixed